static seo friendly build added
