+++ package.json
@@ -17,7 +17,7 @@
   "scripts": {
     "start": "react-scripts start",
     "postbuild": "react-snap",
-    "build": "react-scripts build && cp build/index.html build/404.html",
+    "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
   },
