+++ public/index.html
@@ -4,8 +4,6 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="theme-color" content="#464646">
-    <meta property="og:title" content="ডেভেলপার চিটশিট - বাংলা চিটশিটের ভান্ডার">
-    <meta property="og:description" content="ডেভেলপারদের জন্যে প্রয়োজনীয় সব চিটশিট একসাথে এবং সম্পূর্ণ বাংলা ভাষায়। নিজের মাতৃভাষায় চিটশিটের ভান্ডার সমৃদ্ধ করতে আপনিও যোগ দিন">
     <meta property="og:image" content="https://with.zonayed.me/wp-content/uploads/2018/11/devsonket.png">
     <!--
       manifest.json provides metadata used when your web app is added to the
