+++ src/components/Content.js
@@ -9,7 +9,6 @@ import Page404 from "./404";
 import Footer from "./Footer";
 import { Loader } from "./Loader";
 
-import contributorMap from "../utils/contributorMap";
 import { SingleItem } from "./SingleContent";
 
 const SingleContent = styled.div`
@@ -39,7 +38,6 @@ const SingleContent = styled.div`
 class Content extends Component {
   state = {
     data: "",
-    contributor: "",
     redirect: false
   };
 
@@ -56,36 +54,11 @@ class Content extends Component {
   };
 
   getData = async id => {
-    if (!localStorage.getItem("devCon")) {
-      localStorage.setItem("devCon", "{}");
-    }
     try {
       const { data } = await axios(
         `https://raw.githubusercontent.com/devsonket/devsonket.github.io/develop/data/${id}.json`
       );
-      let contributor;
-      try {
-        const getLocalContributor = JSON.parse(localStorage.getItem("devCon"));
-        const getCurrentId = getLocalContributor[id];
-        const getCurrentTime = Date.now();
-        let compareTime = 1000 * 60 * 60;
-        if (getCurrentId && getCurrentId[1] + compareTime >= getCurrentTime) {
-          contributor = getCurrentId[0];
-        } else {
-          if(navigator.userAgent !== 'ReactSnap') {
-            contributor = await axios(
-              `https://api.github.com/repos/devsonket/devsonket.github.io/commits?path=data/${id}.json`
-            );
-            contributor = contributorMap(contributor.data);
-            let dataForLocalStorage = [contributor, Date.now()];
-            getLocalContributor[id] = dataForLocalStorage;
-            localStorage.setItem("devCon", JSON.stringify(getLocalContributor));
-          }
-        }
-      } catch (e) {
-        contributor = null;
-      }
-      this.setState({ data, contributor });
+      this.setState({ data });
       this.setTitle();
     } catch (e) {
       this.setState({ redirect: true });
@@ -93,10 +66,11 @@ class Content extends Component {
   };
 
   render() {
-    const { data, contributor, redirect } = this.state;
+    const { data, redirect } = this.state;
     const {
       match: { url }
     } = this.props;
+    const id = this.props.match.params.id;
 
     if (redirect && url !== "/404") {
       return <Redirect push to="/404" />;
@@ -115,7 +89,7 @@ class Content extends Component {
         <HeaderContent
           title={data.title}
           description={data.description}
-          contributor={contributor}
+          id={id}
         />
         <Container>
           <SingleContent>
