+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { hydrate, render } from 'react-dom';
 import './index.css';
 import App from './App';
-import registerServiceWorker from './registerServiceWorker';
+// import registerServiceWorker from './registerServiceWorker';
 
 const rootElement = document.getElementById("root");
 if (rootElement.hasChildNodes()) {
@@ -10,4 +10,4 @@ if (rootElement.hasChildNodes()) {
 } else {
   render(<App />, rootElement);
 }
-registerServiceWorker();
+// registerServiceWorker();
